Refactor iOS and Android application entry points to provide all of the "hosting" functionality separate from the Xunit-specific functionality.
